refactor(suite): Move find to suite-desktop-ui
