+++ packages/suite-desktop-ui/package.json
@@ -28,6 +28,7 @@
         "history": "^5.3.0",
         "react": "19.0.0",
         "react-dom": "19.0.0",
+        "react-focus-lock": "^2.13.6",
         "react-redux": "9.2.0",
         "styled-components": "^6.1.19"
     },
