+++ packages/suite-desktop-ui/src/support/FindBar/consts.ts
