+++ packages/suite-desktop-ui/src/support/FindBar/highlight.ts
