+++ packages/suite-desktop-ui/src/support/FindBar/useFindBarShortcuts.ts
