+++ packages/suite/global.d.ts
@@ -40,13 +40,4 @@ declare module 'redux' {
     }
 }
 
-declare global {
-    interface Window {
-        electronFind: {
-            onShow: (callback: () => void) => void;
-            offShow: (callback: () => void) => void;
-        };
-    }
-}
-
 export {};
