+++ packages/suite/src/support/suite/Main.tsx
@@ -4,7 +4,6 @@ import { Router } from 'react-router';
 import type { History } from 'history';
 
 import { FormatterProvider } from '@suite-common/formatters';
-import { isDesktop } from '@trezor/env-utils';
 
 import { useFormattersConfig } from 'src/hooks/suite';
 import Autodetect from 'src/support/suite/Autodetect';
@@ -16,7 +15,6 @@ import Protocol from 'src/support/suite/Protocol';
 import Resize from 'src/support/suite/Resize';
 import { ResponsiveContextProvider } from 'src/support/suite/ResponsiveContext';
 
-import { FindBar } from './FindBar/FindBar';
 import { RouterHandler } from './RouterHandler';
 import { useConnectPopupModals } from './useConnectPopupModals';
 
@@ -31,7 +29,6 @@ export const Main = ({
 }) => {
     useConnectPopupModals();
     const formattersConfig = useFormattersConfig();
-    const isDesktopApp = isDesktop();
 
     return (
         // Todo: Enable when issues are fixed (ReactTruncate & BumpFee)
@@ -51,7 +48,6 @@ export const Main = ({
                                 <FormatterProvider config={formattersConfig}>
                                     {children}
                                 </FormatterProvider>
-                                {isDesktopApp && <FindBar />}
                             </ConnectedIntlProvider>
                         </ErrorBoundary>
                     </ResponsiveContextProvider>
